After this point, you should be able to run gov.pentagon.test.HBaseConnector and see if everything's ok.
You need to make sure you have all the libraries imported, as well the
correct path inside res/preferences.txt . The problem is, I couldn't
commit & ignore this file at the same time, so it's not ignored and it
will not be until you both get this version. At that point we'll ignore
it.